Componentized the following mechanics from the student report page: 1) Session management (I made a general one that all pages can use) 2) Fetch management 3) Filter and memo management
