+++ frontend/src/app/reports/pageStudent.tsx
@@ -1,24 +1,3 @@
-"use client";
-
-import React, {useState, useMemo, useEffect, useRef} from 'react';
-import { apiHandler } from '@/utils/api';
-import { motion, AnimatePresence } from 'framer-motion';
-import Course from '@/components/types/course';
-import Exam from "@/components/types/exam";
-import { getGradeStatus } from '@/app/grades/localComponents/GradeCards';
-import { useSession } from "next-auth/react";
-import { StatusChart } from "./localComponents/StatusChart";
-import ProgressChart from "./localComponents/ProgressChart";
-import { GradeRequirements, GradeStrategy, GradeRequirementsJSON,
-    Report } from "./types/shared"
-import { ExamTable } from "@/app/reports/localComponents/GradeStrategyTable";
-import { Grade } from "@/app/grades/util/types";
-import { StrategyProgressBar } from "./localComponents/StrategyProgressBar";
-import GradeDashboard from "@/app/reports/localComponents/GradeDashboard";
-import { RingSpinner } from "@/components/UI/Spinners";
-import GradeDetermination from "@/app/reports/utils/GradeDetermination";
-import { CourseSelector, allCourse } from "@/components/services/CourseSelector";
-
 /**
  * This is the Student Report page
  * This will give a comprehensive overview
@@ -30,228 +9,98 @@ import { CourseSelector, allCourse } from "@/components/services/CourseSelector"
  * towards any particular grade to know what more
  * is required to receive that grade
  * @author Joshua Summers
+ */
+
+"use client";
+
+import React, { useState, useEffect, useRef } from 'react';
+import { motion, AnimatePresence } from 'framer-motion';
+import { StatusChart } from "./localComponents/StatusChart";
+import ProgressChart from "./localComponents/ProgressChart";
+import { ExamTable } from "@/app/reports/localComponents/GradeStrategyTable";
+import { StrategyProgressBar } from "./localComponents/StrategyProgressBar";
+import GradeDashboard from "@/app/reports/localComponents/GradeDashboard";
+import { RingSpinner } from "@/components/UI/Spinners";
+import { CourseSelector, allCourse } from "@/components/services/CourseSelector";
+import { useSessionData } from "@/hooks/useSessionData";
+import { useFetchData } from "@/app/reports/hooks/useFetchData";
+import { useReportFilters } from "@/app/reports/hooks/useReportFilters";
+
+
+/**
+ * This is the Student Report default page component
  * @constructor
  */
 export function StudentReport() {
-    // These are the session state variables
-    const { data: session, status } = useSession();
-    // Session user information
-    const [sessionReady, setSessionReady] = useState(false);
-    const [userSession, setSession] = useState({
-        id: '',
-        username: '',
-        email: '',
-        accessToken: '',
-    });
+    // This is the Backend API data
+    const BACKEND_API = process.env.NEXT_PUBLIC_BACKEND_API ?? '';
+    // Session hook
+    const { userSession, sessionReady } = useSessionData();
+    // Data Fetch hook
+    const {
+        grades,
+        courses,
+        exams,
+        selectedCourse,
+        setSelectedCourse,
+        loading,
+        fetchAllData,
+        fetchExams
+    } = useFetchData(userSession, BACKEND_API);
+    // Memoized filters hook
+    const {
+        filter,
+        courseFilter,
+        gradeFilter,
+        currentGrade,
+        courseGrades,
+        setGradeFilter,
+        setCourseGrades,
+        updateCourseHandler,
+        filteredGrades,
+        filteredCourse,
+        filteredGradeStrategy,
+    } = useReportFilters(courses, selectedCourse, setSelectedCourse);
 
-    // View data states
-    const [grades, setGrades] = useState<Report[]>([]);
-    const [courseGrades, setCourseGrades] = useState<Report[]>([]);
-    const [courses, setCourses] = useState<Course[]>([]);
-    const [course, setCourse] = useState<Course>();
-    const [exams, setExams] = useState<Exam[]>([]);
-    const [gradeStrategy, setGradeStrategy] = useState<GradeStrategy>();
-    const [currentGrade, setCurrentGrade] = useState<String>('A');
     // Toggle states
-    const [loading, setLoading] = useState(true);
     const [refreshTrigger, setRefreshTrigger] = useState(0);
 
     // Reference to control React double render of useEffect
     const hasFetched = useRef(false);
 
-    // Filter states
-    const [filter, setFilter] = useState<'all' | 'passed' | 'failed' | 'pending'>('all');
-    const [courseFilter, setCourseFilter] = useState<string>('');
-    const [gradeFilter, setGradeFilter] = useState<'A' | 'B' | 'C' | 'D' | 'F'>('A');
-
-    const BACKEND_API = process.env.NEXT_PUBLIC_BACKEND_API;
-
-    // Load the specific grade strategy details
-    const loadGradeStrategy = (strategy: GradeStrategy) => {
-        // Define the grade letter strategy
-        let newStrategy: GradeStrategy = {
-            total: strategy?.total || 0,
-            requiredA: strategy?.requiredA || 0,
-            optional: strategy?.optional || [],
-            allOptional: strategy?.allOptional || false
-        }
-        // Return the new Grade Strategy
-        return newStrategy;
-    }
-
-    // Grade Calculation based on course strategy
-    const calculateCurrentGrade = (passed: number, passedAs: number,
-                                   requirements: GradeRequirements) => {
-        if (passed >= requirements.A.total && passedAs >= requirements.A.requiredA) return 'A';
-        if (passed >= requirements.B.total && passedAs >= requirements.B.requiredA) return 'B';
-        if (passed >= requirements.C.total && passedAs >= requirements.C.requiredA) return 'C';
-        if (passed >= requirements.D.total && passedAs >= requirements.D.requiredA) return 'D';
-        return 'F';
-    }
-
-
-    // Memoized computations (IMPROVED)
-    const filteredGrades = useMemo(() => {
-        if (!courseGrades || courseGrades.length === 0) return [];
-        if (filter === 'all') return courseGrades;
-        return courseGrades.filter(grade => getGradeStatus(grade as Grade) === filter);
-    }, [courseGrades, filter]);
-
-    const filteredCourses = useMemo(() => {
-        if (!courses || courses.length === 0) return [];
-        return courses.filter(course => course.courseName === courseFilter);
-    }, [courses, courseFilter]);
-
-    // FIXED: More efficient grade strategy calculation
-    const filteredGradeStrategy = useMemo(() => {
-        if (!filteredCourses?.[0]?.gradeStrategy) return null;
-
-        try {
-            console.log('Inside the filtered grade strategy memo');
-            const strategy: GradeRequirementsJSON = JSON.parse(filteredCourses[0].gradeStrategy);
-            let selectedStrategy: GradeStrategy;
-
-            switch (gradeFilter) {
-                case 'A': selectedStrategy = strategy.GradeA; break;
-                case 'B': selectedStrategy = strategy.GradeB; break;
-                case 'C': selectedStrategy = strategy.GradeC; break;
-                case 'D': selectedStrategy = strategy.GradeD; break;
-                default: selectedStrategy = strategy.GradeF;
-            }
-
-            return {
-                required: strategy.requiredExams,
-                optional: strategy.optionalExams,
-                strategy: selectedStrategy
-            };
-        } catch (error) {
-            console.error('Error parsing grade strategy:', error);
-            return null;
-        }
-    }, [filteredCourses, gradeFilter]);
-
-    // FIXED: Memoized grade requirements calculation
-    const gradeRequirements = useMemo(() => {
-        if (!filteredCourses?.[0]?.gradeStrategy) return null;
-
-        try {
-            console.log('Inside the grade requirements memo');
-            const strategyJSON: GradeRequirementsJSON = JSON.parse(filteredCourses[0].gradeStrategy);
-            return {
-                A: loadGradeStrategy(strategyJSON.GradeA),
-                B: loadGradeStrategy(strategyJSON.GradeB),
-                C: loadGradeStrategy(strategyJSON.GradeC),
-                D: loadGradeStrategy(strategyJSON.GradeD),
-                F: loadGradeStrategy(strategyJSON.GradeF)
-            };
-        } catch (error) {
-            console.error('Error calculating grade requirements:', error);
-            return null;
-        }
-    }, [filteredCourses]);
-
-    // FIXED: Memoized current grade calculation
-    const calculatedCurrentGrade = useMemo(() => {
-
-        console.log('CALCULATED_CURRENT_GRADE useMemo RUNNING');
-        console.log('gradeRequirements:', gradeRequirements);
-        console.log('filteredGrades:', filteredGrades);
-        console.log('filteredGrades length:', filteredGrades?.length);
-        console.log('grades:', grades);
-
-        // // If no grades requirements, return F
-        // if (!gradeRequirements) return 'F';
-        // If we have grades to make a grade determination on
-        if (gradeRequirements && filteredGrades && filteredGrades.length > 0)
-            return GradeDetermination(filteredGrades, gradeRequirements);
-        else if (!gradeRequirements && filteredGrades && filteredGrades.length > 0) {
-            console.log('No Strategy Grade Determination');
-            return GradeDetermination(filteredGrades);
-        }
-        // If neither of those exist, return F
-        else return 'F'
-    }, [filteredGrades, gradeRequirements, gradeFilter]);
-
-    // EFFECT 1: Initial data loading (IMPROVED)
+    /**
+     * This useEffect fetches the data once
+     * the session has been hydrated
+     */
     useEffect(() => {
-        // Exit if session not ready
-        if (!sessionReady) return;
-        // Wrapper for async function
-        const fetchData = async () => {
-            if (hasFetched.current) return;
+        if (!sessionReady || hasFetched.current) return;
+
+        const hydrateData = async () => {
             hasFetched.current = true;
-            setLoading(true);
+            await fetchAllData();
+        }
+        // Call the async handler
+        hydrateData();
+    }, [sessionReady, fetchAllData, refreshTrigger]);
 
-            try {
-                const examsRaw = await fetchGrades();
-                const coursesData = await fetchCourses(examsRaw);
-
-                if (coursesData.length > 0) {
-                    // Set initial course filter
-                    setCourseFilter(coursesData[0].courseName);
-                }
-            } catch (error) {
-                console.error('Error fetching initial data:', error);
-            } finally {
-                setLoading(false);
-            }
-        };
-
-        fetchData();
-    }, [sessionReady, userSession.id, hasFetched, refreshTrigger]);
-
-    // EFFECT 2: Course-specific data
+    /**
+     * This useEffect ties the fetch hook states
+     * and the report filter states together
+     */
     useEffect(() => {
-        if (!filteredCourses?.[0] || loading) return;
+        if (!selectedCourse || loading) return;
 
-        const updateCourseData = async () => {
+        const updateCourseExams = async () => {
             // Update course grades
             const reducedGrades = grades.filter(grade =>
                 grade.courseName === courseFilter);
             setCourseGrades(reducedGrades);
-
             // Fetch exams for this course
-            await fetchExams(filteredCourses[0]);
-        };
-
-        updateCourseData();
-    }, [filteredCourses?.[0]?.courseId, courseFilter, loading]);
-
-    // EFFECT 3: Grade strategy updates
-    useEffect(() => {
-        if (filteredGradeStrategy) {
-            console.log('Inside the filtered grade strategy useAffect');
-            setGradeStrategy(filteredGradeStrategy.strategy);
+            await fetchExams(selectedCourse);
         }
-    }, [filteredGradeStrategy]);
-
-    // EFFECT 4: Current grade updates
-    useEffect(() => {
-        console.log('Calculate current grade:');
-        console.log(calculatedCurrentGrade)
-        setCurrentGrade(calculatedCurrentGrade);
-    }, [calculatedCurrentGrade]);
-
-    /**
-     * useAffects for session hydration
-     */
-    // EFFECT 5: Session management (separate from data fetching)
-    useEffect(() => {
-        let id = '';
-        if (status !== 'authenticated' || !session || hasFetched.current) return;
-        // Hydrate session information
-        if (session) {
-            const newUserSession = {
-                id: session?.user.id?.toString() || '',
-                username: session?.user.username || '',
-                email: session?.user.email || '',
-                accessToken: session?.user.accessToken || '',
-            };
-
-            setSession(newUserSession);
-            setSessionReady(newUserSession.id !== "");
-        }
-    }, [session, status]);
+        // Call the async handler
+        updateCourseExams();
+    }, [selectedCourse?.courseId, fetchExams]);
 
     // Motion Container Variant definitions
     const containerVariants = {
@@ -264,221 +113,20 @@ export function StudentReport() {
         }
     };
 
-    /**
-     * Average Grade Determination
-     * Implementation for getting the average of the grades
-     */
-    const avgScore = (grades: Report[]) => {
-        // Variables
-        let counter = 0;
-        let finalGrade;
-        // Process each element (async)
-        grades.forEach((grade: Report) => {
-            if (grade.examScore == 'A') counter += 5;
-            else if (grade.examScore == 'B') counter += 4;
-            else if (grade.examScore == 'C') counter += 3;
-            else if (grade.examScore == 'D') counter += 2;
-            else if (grade.examScore == 'F') counter += 1;
-        });
-
-        // Get the average of the scores
-        const avgGrade = Math.round(counter / grades.length);
-        switch (avgGrade) {
-            case 5:
-                finalGrade = 'A';
-                break;
-            case 4:
-                finalGrade = 'B'
-                break;
-            case 3:
-                finalGrade = 'C';
-                break;
-            case 2:
-                finalGrade = 'D';
-                break;
-            default:
-                finalGrade = 'F';
-        }
-        // Return the average letter grade for all grades
-        return finalGrade;
-    }
-
-    /**
-     * Fetch Grades
-     * Implementation for general API handler
-     */
-    async function fetchGrades() {
-        console.log('Fetching data for student report page: Grade data');
-        setLoading(true);
-        let examsRaw: Report[] = [];
-
-        try {
-            const res = await apiHandler(
-                undefined,
-                'GET',
-                `api/exam/result/grades/${userSession.id}`,
-                `${BACKEND_API}`,
-                userSession.accessToken
-            );
-            console.log(res);
-
-            if (res instanceof Error || (res && res.error)) {
-                console.error('Error fetching grades:', res.error);
-            } else {
-                // Get all the grades
-                examsRaw = res.grades || res || []; // Once grabbed, it is gone
-                // Ensure each grade has a proper status
-                examsRaw = examsRaw.map(grade => ({
-                    ...grade,
-                    status: getGradeStatus(grade as Grade)
-                }));
-            }
-        } catch (error) {
-            console.error('Error fetching student grades:', error as string);
-        } finally {
-            if (examsRaw.length === 0) {
-                setGrades([]);
-            } else {
-                setGrades(examsRaw);
-            }
-            return examsRaw;
-        }
-    }
-
-    /**
-     * Fetch Courses
-     * Implementation for general API handler
-     */
-    async function fetchCourses(examsRaw: Report[]) {
-        console.log('Fetching data for student report page: Course data');
-        // setLoading(true);
-        console.log(examsRaw);
-        // Determine courses based on exams
-        let courseIdList = examsRaw.reduce((unique: string[], grade) => {
-            const courseId = grade.courseId;
-            if (courseId && !unique.includes(courseId.toString())) {
-                unique.push(courseId.toString());
-            }
-            return unique;
-        }, []);
-        console.log('Course Id List');
-        console.log(courseIdList);
-        // Courses List
-        let coursesData: Course[] = [];
-        // Exception Wrapper for API handler
-        try {
-            // Iterate through the course Ids
-            for (const courseId of courseIdList) {
-                const res = await apiHandler(
-                    undefined,
-                    'GET',
-                    `api/course/${courseId}`,
-                    `${BACKEND_API}`,
-                    userSession.accessToken
-                );
-                console.log('Course API response');
-                console.log(res);
-
-                if (res instanceof Error || (res && res.error)) {
-                    console.error('Error fetching course:', res.error);
-                } else {
-                    // Store the course in the list
-                    coursesData.push(res.course || res);
-                }
-            }
-
-            console.log('This is the courses data:');
-            console.log(coursesData);
-
-        } catch (error) {
-            console.error('Error fetching student courses:', error as string);
-            // setCourses([]);
-        } finally {
-            if (coursesData.length !== 0) {
-                // Time to update states
-                // First course retried as default
-                setCourseFilter(coursesData[0].courseName);
-                setCourses(coursesData);
-            } else {
-                setCourses([]);
-            }
-            // Now we are done loading data
-            // setLoading(false);
-            return coursesData;
-        }
-    }
-
-    /**
-     * Fetch Exams
-     * Implementation for general API handler
-     */
-    async function fetchExams(course: Course) {
-        console.log('Fetching data for exams: Exam data');
-        console.log(course);
-        let examsRaw: Exam[] = [];
-
-        try {
-            const res = await apiHandler(
-                undefined,
-                'GET',
-                `api/exam/course/${course.courseId}`,
-                `${BACKEND_API}`,
-                userSession.accessToken
-            );
-            console.log(res);
-
-            if (res instanceof Error || (res && res.error)) {
-                console.error('Error fetching exams for exams:', res.error);
-            } else {
-                // Get all the exams
-                examsRaw = res.exams || res || []; // Once grabbed, it is gone
-            }
-        } catch (error) {
-            console.error('Error fetching course exams:', error as string);
-        } finally {
-            if (examsRaw.length === 0) {
-                setExams([]);
-            } else {
-                setExams(examsRaw);
-            }
-            // setLoading(false);
-            console.log('Fetched Exams');
-            console.log(examsRaw);
-            return examsRaw;
-        }
-    }
-
-    // Handle Course Updates from Course Selector Components
-    const updateCourseHandle = async (courseId: string) => {
-        // Turn the string into an integer
-        let courseIdInt = parseInt(courseId);
-        // First case is the default All course
-        if (courseIdInt === -1) {
-            setCourseFilter('all')
-            setCourse(allCourse);
-        }
-        // This is the
-        else {
-            let reduced = courses.find(course =>
-                course.courseId === courseIdInt);
-            console.log(reduced);
-            if (reduced) {
-                setCourseFilter(reduced.courseName);
-                setCourse(reduced);
-            }
-        }
-    }
-
     return (
         <div className="px-4 py-4">
             {/*This is the course header*/}
             <div className="max-w-5xl mx-auto flex justify-between items-center mb-6">
                 {/*This is the Course Selection button*/}
-                {session ?
+                {userSession ?
                     (
                         <React.Fragment>
-                            <h2 className="text-xl font-semibold">{session?.user?.name}'s Report</h2>
-                            {loading ? (<React.Fragment/>) : filteredCourses.length === 0
+                            <h2 className="text-xl font-semibold">{userSession.name
+                                ? userSession.name + "'s Report"
+                                : ''
+                            }
+                            </h2>
+                            {loading ? (<React.Fragment/>) : !selectedCourse
                                 ? (
                                     <div>
                                         <p>Student has no courses</p>
@@ -487,9 +135,9 @@ export function StudentReport() {
                                 : courses && courses.length > 0 && (
                                     <CourseSelector
                                         courses={courses}
-                                        selectedCourseId={course?.courseId}
+                                        selectedCourseId={selectedCourse.courseId}
                                         onCourseChange={(e) => {
-                                            updateCourseHandle(e.target.value);
+                                            updateCourseHandler(e.target.value);
                                             console.log(filter);
                                         }}
                                     />
@@ -524,7 +172,6 @@ export function StudentReport() {
                                         ) :
                                         (
                                             <motion.div
-                                                // key={`${selectedClass}-${filter}`}
                                                 // variants={containerVariants}
                                                 initial="hidden"
                                                 animate="visible"
@@ -547,7 +194,6 @@ export function StudentReport() {
                                         ) :
                                         (
                                             <motion.div
-                                                // key={`${selectedClass}-${filter}`}
                                                 // variants={containerVariants}
                                                 initial="hidden"
                                                 animate="visible"
@@ -556,7 +202,7 @@ export function StudentReport() {
                                                 <ProgressChart
                                                     key={`progress-chart-${courseFilter}`}
                                                     exams={filteredGrades}
-                                                    course={filteredCourses[0]}
+                                                    course={selectedCourse}
                                                     currentGrade={currentGrade}
                                                 />
                                             </motion.div>
@@ -630,12 +276,6 @@ export function StudentReport() {
                                 >
                                     F
                                 </button>
-                                {/*<button*/}
-                                {/*    className={`px-4 py-2 rounded-lg text-sm font-medium transition-colors ${filter === 'pending' ? 'bg-crimson text-mentat-gold-700' : 'bg-crimson text-mentat-gold hover:bg-crimson-700'}`}*/}
-                                {/*    onClick={() => setFilter('pending')}*/}
-                                {/*>*/}
-                                {/*    Pending*/}
-                                {/*</button>*/}
                             </div>
                         </div>
                         <hr className="border-crimson border-1 my-4"></hr>
@@ -686,21 +326,14 @@ export function StudentReport() {
                                     <h2 className="text-xl font-semibold">
                                         Grade {gradeFilter} Course Requirements
                                     </h2>
-                                    {!filteredGradeStrategy ? (
-                                        <React.Fragment/>
-                                    ) : (
-                                        <span className="text-sm text-mentat-gold/80 italic">
-                                            {filteredGradeStrategy.strategy.total} exams required with score greater than C <br/>
-                                            {filteredGradeStrategy.strategy.requiredA} exams with an A score required
-                                        </span>
-                                    )}
+                                    <span className="text-sm text-mentat-gold/80 italic">
+                                        {filteredGradeStrategy.strategy.total} exams required with score greater than C <br/>
+                                        {filteredGradeStrategy.strategy.requiredA} exams with an A score required
+                                    </span>
                                 </div>
                                 {/*New Grade Strategy Chart*/}
                                 <div className="pt-1 mt-2 bg-card-color shadow-sm shadow-crimson-700">
-                                    {!filteredGradeStrategy && !filteredCourses
-                                    && !exams && !courseGrades ? (
-                                        <React.Fragment/>
-                                    ) : (
+                                    {filteredCourse && exams && courseGrades && (
                                         <AnimatePresence mode="wait">
                                             <ExamTable
                                                 grades={courseGrades}
@@ -708,76 +341,17 @@ export function StudentReport() {
                                                 required={filteredGradeStrategy.required}
                                                 optional={filteredGradeStrategy.optional}
                                                 exams={exams}
-                                                course={filteredCourses[0]}
+                                                course={filteredCourse}
                                                 currentGrade={currentGrade}
                                             />
                                         </AnimatePresence>
                                     )}
                                 </div>
-
-                                {/*<div className="max-h-[600px] pt-1">*/}
-                                {/*    <AnimatePresence mode="wait">*/}
-                                {/*        {filteredGrades.length > 0 ? (*/}
-                                {/*            <motion.div*/}
-                                {/*                key={`${selectedClass}-${filter}`}*/}
-                                {/*                variants={containerVariants}*/}
-                                {/*                initial="hidden"*/}
-                                {/*                animate="visible"*/}
-                                {/*                className="space-y-2 mb-2"*/}
-                                {/*            >*/}
-                                {/*                {filteredGrades.map((grade) => (*/}
-                                {/*                    <GradeCardExtended*/}
-                                {/*                        key={grade.exam_id}*/}
-                                {/*                        grade={grade}*/}
-                                {/*                        index={0}*/}
-                                {/*                        onclick={(e) => loadGradeDetails(grade, e)}*/}
-                                {/*                    />*/}
-                                {/*                ))}*/}
-                                {/*            </motion.div>*/}
-                                {/*        ) : loading === true ? (*/}
-                                {/*            <motion.div*/}
-                                {/*                variants={containerVariants}*/}
-                                {/*                initial={{ opacity: 0 }}*/}
-                                {/*                animate={{ opacity: 1 }}*/}
-                                {/*                className="text-center py-12"*/}
-                                {/*            >*/}
-                                {/*                Loading Grades...*/}
-                                {/*            </motion.div>*/}
-                                {/*        ) : (*/}
-                                {/*            <motion.div*/}
-                                {/*                variants={containerVariants}*/}
-                                {/*                initial={{ opacity: 0 }}*/}
-                                {/*                animate={{ opacity: 1 }}*/}
-                                {/*                className="text-center py-12"*/}
-                                {/*            >*/}
-                                {/*                No grades found for the selected filters*/}
-                                {/*            </motion.div>*/}
-                                {/*        )}*/}
-                                {/*    </AnimatePresence>*/}
-                                {/*</div>*/}
                             </motion.div>
                         )}
                     </div>
                 </motion.div>
             </div>
-
-
-            {/*/!* Exam Action Modal *!/*/}
-            {/*<Modal*/}
-            {/*    isOpen={isExamModalOpen}*/}
-            {/*    onClose={() => setIsExamModalOpen(false)}*/}
-            {/*    title="Alter Scheduled Exam"*/}
-            {/*>*/}
-            {/*    <ExamActionsComponent*/}
-            {/*        examResult={examResult}*/}
-            {/*        cancelAction={() => {*/}
-            {/*            setIsExamModalOpen(false);*/}
-            {/*            // Trigger refresh when modal closes*/}
-            {/*            setRefreshTrigger(prev => prev + 1);*/}
-            {/*        }}*/}
-            {/*    />*/}
-            {/*</Modal>*/}
-
         </div>
     );
 }
